Pare down Travis CI build matrix to save on build credits
This commit shaves off 10 permutations from the build matrix, because
this setup:

    (windows, osx) X (versions) X (modules off, modules on)

Was quickly depleting my Travis credits. We shave the main matrix down
to:

    (linux) X (versions) X (modules off, modules on)

which is currently 6 builds, and add an additonal 4 ad hoc build jobs
on the GoLang master branch. The ad hoc jobs should hopefully cover
Windows and OSX "enough" to ensure the code work well there.