Don't cache execution plan (allow handlers to change it)
This change is driven by development on the AWS X-Ray plugin for httpx,
github.com/gogama/aws-xray-httpx. The plugin needs to be able to change
the plan context (which is only possible by shallow copying the whole
plan). The reason it needs to be able to do this is so the individual
attempt requests have a context that inherits from the context which
contains the X-Ray "parent" subsegment representing the whole execution.
Without this, the "child" attempt subsegments don't get attached to the
execution subsegment and they just disappear.

Apart from being driven by a real-world use case, this change makes
sense from a consistency perspective. The handleCheckpoint() code
already allows event handlers to change the current attempt request
on the execution, or the response, so why not also the plan?