Increase robustness of goroutine panic handling, fix flaky Close test
The overall robustness improvements mean the client should handle panics
and unexpected conditions in the goroutines it has started in a
reasonable way, clean up, and communicate the panics to the user on the
main (Do invoking) goroutine. Yay.

The specific change which should fix the flaky Close tests is moving the
Body.Close() call out of the defer statement and into the main body of
sendAndReadBody(), before signalling the main goroutine to process via
`as.es.signal <- as`. The reason I believe the defer statement was
causing the test to be flaky is that occasionally the main goroutine
would unblock, complete, return control to the test case, and the test
case would assert expectations on the mock ReadCloser, all before the
sendAndReadBody() goroutine's deferred funtion ran.