Improve reliability of plan timeout detection, fixing flaky test
This commit fixes the flaky test case "TestClient/attempt timeout/from
plan deadline", which was failing around 1% of the time because the
robust HTTP client was failing to trigger the AfterPlanTimeout event
as expected.

I'm not 100% clear on why this was happening other than that the effect
of the plan deadline on the attempt's request sub-context was being
detected in the Go HTTP I/O layer, causing the request to fail with a
timeout, before the plan context's deadline timer had a chance to run,
or at any rate before the goroutine that handles the deadline timer
event had a chance to run and cancel the plan context.

This change makes plan timeout detection more robust by explicitly
testing the plan context for deadline expiry if the plan context hasn't
been updated with an error. The added robustness fixes the flaky test
case.

Testing:

$ go test run "TestClient/attempt timeout/from plan deadline" \
    -failfast -count=3000

Previously it was not possible to run nearly this many test reps
without getting an error.