Fix flaky plan timeout during retry wait test
This commit should fix the testClientRetryPlanTimeout test, which is
intended to test the case where the plan times out (or is cancelled)
while the client is sitting in a retry wait between attempts.

The cause of the flakiness is that in rare cases the plan timeout may
be detected before the retry Decide() method is called, so that
neither Decide() nor Wait() are exercised on the retry policy mock
even though all aspects of the test behave correctly.

This fix is to switch those mock expectations from .Once() to .Maybe().