Classify unexpected end-of-file as transient by default
The result of this change is that retry.TransientErr will indicate a
retry if an unexpected EOF is seen when writing to or reading from
the connection, regardless of whether the EOF is seen before reading
the response headers or afterward.

As usual, clients interacting with non-idempotent web services should
think carefully about what type of retry strategy is the right one
for them.